refactor(api): collapse altrep_impl declarative macro variants (Path B from #682)
`miniextendr-api/src/altrep_impl.rs` was 2591 lines of declarative-macro
plumbing for 7 ALTREP family `impl_alt<family>_from_data!` macros, with
~10 arms per family covering combinatorial knob subsets, plus reverse-
spelling aliases and a parallel materializing path.

This PR consolidates the macro surface and splits the file along its
natural layers. Public derive surface (`#[derive(AltrepInteger)]` etc.)
is unchanged.

## Macro consolidation

- Merge `__impl_altrep_base!` and `__impl_altrep_base_with_serialize!`
  into one macro with an optional `with_serialize` token (4 arms covering
  default-guard / explicit-guard × with-or-without serialize). Update the
  three call sites in `altrep_derive.rs`.

- Delete the `materializing_dataptr` and `materializing_dataptr, serialize`
  arms from all 6 numeric/string/logical families. They were byte-identical
  to the default `($ty)` and `($ty, serialize)` arms — both routed through
  `__impl_altvec_<family>_dataptr!` which provides a trivial
  `AltrepDataptr<T>` returning `None`, falling through to data2
  materialization. The only call site was `impl_builtin_altrep_family!`'s
  `materializing` mode, migrated to use the default arms.

- Delete reverse-spelling aliases (`serialize, dataptr` and
  `serialize, subset`) from all 6 numeric/string/logical families. The
  proc-macro derive only emits canonical alphabetical order; no internal
  call site uses the reverse spelling.

Result: 10-arm families collapsed to 6 arms each (6 × 4 deleted = 24 arms
removed from `altrep_impl.rs`).

## FFI _unchecked normalisation

Standardise the two raw `Rf_protect(Rf_allocVector(STRSXP, n))` sites
(`__impl_altvec_string_dataptr!`, `__impl_altstring_methods!::elt`) onto
`Rf_protect_unchecked` / `Rf_allocVector_unchecked` /
`Rf_unprotect_unchecked` uniformly. Both run inside ALTREP callbacks
where `_unchecked` variants are MXL301-permitted. Matches the existing
discipline in `__impl_altrep_base!`'s `with_serialize` arm.

## File split (no `mod.rs` per CLAUDE.md)

`altrep_impl.rs` (71 lines) — public helpers (`checked_mkchar`,
`altrep_region_buf`) + `mod` declarations for the four submodules.

`altrep_impl/macros.rs` (1256 lines) — all `#[macro_export]`
declarative macros (`impl_alt<family>_from_data!` + their `__impl_*`
helpers + shared building blocks). The bulk of the file.

`altrep_impl/builtins.rs` (590 lines) — the crate-private
`impl_builtin_altrep_family!` meta-macro + `impl_register_altrep_builtin!`
+ all `Vec<T>` / `Box<[T]>` / `Cow<T>` / `Range<T>` invocations and
`RegisterAltrep` impls.

`altrep_impl/arrays.rs` (255 lines) — const-generic `[T; N]` ALTREP
impls. Uses local (non-exported) `impl_altrep_array_numeric!`.

`altrep_impl/static_slices.rs` (427 lines) — hand-written `&'static [T]`
impls (different writable-assert pattern; not macro-reachable).

Path (a) from #682 (proc-macro rewrite of the declarative layer) is
deferred to a follow-up issue, as is the parallel `altrep_data/builtins.rs`
audit.
